+++ src/DBClient/DBClient.cpp
@@ -229,9 +229,9 @@ std::string DBClient::getDefinesReactionSymbol(std::string _idSubst)
 
 void DBClient::setReactantsFollowingIncomingTakesEdges(std::string _id, Reaction &reaction)
 {
-    std::map<std::string, int> map;
+    std::map<std::string, double> map;
     string kbuf;
-    int stoi_coeff;
+    double stoi_coeff;
     bson record;
     string qrJson = "{'_type': 'edge', '_label': 'takes', '_inV': '";
     qrJson += _id;
@@ -245,7 +245,7 @@ void DBClient::setReactantsFollowingIncomingTakesEdges(std::string _id, Reaction
     {
         jsonToBson(&record, _resultDataEdge[i]);
         bsonio::bson_to_key( record.data, "properties.stoi_coeff", kbuf );
-        stoi_coeff = atoi(kbuf.c_str());
+        stoi_coeff = atof(kbuf.c_str());
 
         bsonio::bson_to_key( record.data, "_outV", kbuf );
         qrJson = "{ \"_id\" : \""+kbuf+ "\"}";
@@ -257,7 +257,7 @@ void DBClient::setReactantsFollowingIncomingTakesEdges(std::string _id, Reaction
             bsonio::bson_to_key( record.data, "properties.symbol", kbuf );
         }
 
-        map.insert(std::pair<std::string,int>(kbuf,stoi_coeff));
+        map.insert(std::pair<std::string,double>(kbuf,stoi_coeff));
     }
 
     reaction.setReactants(map);
